+++ src/js/options.ts
@@ -39,5 +39,37 @@ export interface CordovaOptions
   *
   * @default true
   * */
+  enableWatchdogTerminationTracking?: boolean;
+
+  /**
+  * Enables Out of Memory Tracking for iOS and macCatalyst.
+  * See the following link for more information and possible restrictions:
+  * https://docs.sentry.io/platforms/apple/guides/ios/configuration/out-of-memory/
+  *
+  * @default true
+  * @deprecated The method will be removed on a major update, instead, use enableWatchdogTerminationTracking for the same result.
+  * */
   enableOutOfMemoryTracking?: boolean;
+
+ /**
+  * When enabled, the SDK tracks when the application stops responding for a specific amount of
+  * time defined by the `appHangTimeoutInterval` option.
+  *
+  * iOS only
+  *
+  * @default true
+  */
+ enableAppHangTracking?: boolean;
+
+ /**
+  * The minimum amount of time an app should be unresponsive to be classified as an App Hanging.
+  * The actual amount may be a little longer.
+  * Avoid using values lower than 100ms, which may cause a lot of app hangs events being transmitted.
+  * Value should be in seconds.
+  *
+  * iOS only
+  *
+  * @default 2
+  */
+ appHangTimeoutInterval?: number;
 }