+++ CHANGELOG.md
@@ -1,5 +1,39 @@
 # Changelog
 
+## 1.3.0
+
+### Features
+
+- Add options for iOS: enableAppHangTracking and appHangTimeoutInterval, allowing users to define the App hang timeout or completly disabling it. ([#338](https://github.com/getsentry/sentry-cordova/pull/338))
+
+### Dependencies
+
+- Bump `sentry-wizard` to 3.21.0 ([#544](https://github.com/getsentry/sentry-wizard/pull/544))
+- Bump Android SDK from v6.11.0 to v7.6.0 ([#336](https://github.com/getsentry/sentry-capacitor/pull/336))
+  - [changelog](https://github.com/getsentry/sentry-java/blob/main/CHANGELOG.md#760)
+  - [diff](https://github.com/getsentry/sentry-java/compare/6.11.0...7.6.0)
+- build(ios): Bump `sentry-cocoa` to 8.21.0 ([#337](https://github.com/getsentry/sentry-cordova/pull/337))
+  - [changelog](https://github.com/getsentry/sentry-cocoa/releases/tag/8.21.0)
+  - [diff](https://github.com/getsentry/sentry-cocoa/compare/8.13.1...8.21.0)
+
+## 1.2.0
+
+### Fixes
+
+-  Fix iOS not bundling ([#316](https://github.com/getsentry/sentry-cordova/pull/316))
+
+### Dependencies
+
+- Bump `sentry-wizard` to 3.16.1 ([#481](https://github.com/getsentry/sentry-wizard/pull/481))
+- build(ios): Bump `sentry-cocoa` to 8.13.1 ([#316](https://github.com/getsentry/sentry-cordova/pull/316))
+  - [changelog](https://github.com/getsentry/sentry-cocoa/releases/tag/8.13.1)
+  - [diff](https://github.com/getsentry/sentry-cocoa/compare/7.27.1...8.13.1)
+
+### Warning, breaking changes
+
+- The option `enableOutOfMemoryTracking` is now `enableWatchdogTerminationTracking`. The previous name will keep working but will be removed on a major version.
+- Bump minimum supported OS versions to macOS 10.13, iOS 11.
+
 ## 1.1.5
 
 ### Fixes